fix disappearing card bug
Whenever you played a card to the table and had a second card of the same color and value in your hand, the matching card in your hand would disappear. This was because the temp table wasn't being cleared when the cards were played to the main table, so the hand update event was still removing the temporary table cards from your hand before displaying it.

So this commit fixes the bug by clearing the temporary table whenever the server responds that you've successfully completed the level.